+++ src/lua/functions/creatures/monster/loot_functions.hpp
@@ -23,7 +23,6 @@ public:
 
 private:
 	static int luaCreateLoot(lua_State* L);
-	static int luaDeleteLoot(lua_State* L);
 	static int luaLootSetId(lua_State* L);
 	static int luaLootSetIdFromName(lua_State* L);
 	static int luaLootSetMinCount(lua_State* L);