+++ src/lua/functions/creatures/monster/monster_type_functions.cpp
@@ -63,6 +63,7 @@ void MonsterTypeFunctions::init(lua_State* L) {
 	Lua::registerMethod(L, "MonsterType", "maxHealth", MonsterTypeFunctions::luaMonsterTypeMaxHealth);
 	Lua::registerMethod(L, "MonsterType", "runHealth", MonsterTypeFunctions::luaMonsterTypeRunHealth);
 	Lua::registerMethod(L, "MonsterType", "experience", MonsterTypeFunctions::luaMonsterTypeExperience);
+	Lua::registerMethod(L, "MonsterType", "skull", MonsterTypeFunctions::luaMonsterTypeSkull);
 
 	Lua::registerMethod(L, "MonsterType", "faction", MonsterTypeFunctions::luaMonsterTypeFaction);
 	Lua::registerMethod(L, "MonsterType", "enemyFactions", MonsterTypeFunctions::luaMonsterTypeEnemyFactions);
@@ -572,6 +573,26 @@ int MonsterTypeFunctions::luaMonsterTypeExperience(lua_State* L) {
 			lua_pushnumber(L, monsterType->info.experience);
 		} else {
 			monsterType->info.experience = Lua::getNumber<uint64_t>(L, 2);
+			Lua::pushBoolean(L, true);
+		}
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int MonsterTypeFunctions::luaMonsterTypeSkull(lua_State* L) {
+	// get: monsterType:skull() set: monsterType:skull(str/constant)
+	const auto &monsterType = Lua::getUserdataShared<MonsterType>(L, 1);
+	if (monsterType) {
+		if (lua_gettop(L) == 1) {
+			lua_pushnumber(L, monsterType->info.skull);
+		} else {
+			if (Lua::isNumber(L, 2)) {
+				monsterType->info.skull = Lua::getNumber<Skulls_t>(L, 2);
+			} else {
+				monsterType->info.skull = getSkullType(Lua::getString(L, 2));
+			}
 			Lua::pushBoolean(L, true);
 		}
 	} else {