+++ markdowns/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Added a new feature to unlock cosmetics in `config.lua`: `unlockAllOutfits`, `unlockAllMounts`, and `unlockAllFamiliars`. ([Tryller](https://github.com/jprzimba))
 - Added a new attributies to `vocations.xml`: `wandRodDamage`, `absorb`. ([Tryller](https://github.com/jprzimba))
+- Add new lua functions: `npc:getMasterPos()`, `player:getOpenContainers()`, `player:isNearDepotBox()`, `vocation:getAbsorbPercent(combatType)`. `vocation:increaseAbsorbPercent(combatType, value)`, `house:isInAccessList(player, listId)`, `house:getAccessLevel(player)`, `house:getOwnerName()`, `house:getOwnerAccountId()`, `itemType:getGroup()`, `itemType:getRotateId()`, `itemType:getWareId()`, `itemType:getCorpseType()`, `itemType:getClassification()`, `itemType:isWrapKit()`, `itemType:isSpellbook()`, `itemType:getAbilities()`, `monsterType:skull()`, `monsterType:skull(str/constant)`, `party:getMinLevel()`, `party:getMaxLevel()`. ([Tryller](https://github.com/jprzimba))
 
 ## Added files
 - NOTTING
@@ -26,6 +27,7 @@
 
 ## Modified files
 
+- crystalserver.exe
 - config.lua
 - data/scripts/actions/items/potions.lua
 - data/XML/vocations.xml